Remove unnecessary ignore flag

+++ onnx/defs/function.cc
@@ -65,7 +65,7 @@ Status FunctionBuilderRegistry::GetFunctions(
         function_proto->since_version() < version_range.first) {
       fail_check("Invalid function version in '", function_proto->name(), "'");
     }
-    op_set.insert({func_builder.GetDomain(), function_proto->since_version()});
+    op_set.insert({func_builder.GetDomain(), (int)function_proto->since_version()});
     ctx.set_opset_imports(op_set);
     ctx.set_is_main_graph(false);
     LexicalScopeContext lex_ctx;
